refactor: update address validator to use BlockchainNetwork interface
