+++ lib/common/src/interfaces/blockchain.network.ts
@@ -4,4 +4,5 @@ export const BlockchainNetwork = {
   ...Network,
 } as const;
 
-export type BlockchainNetwork = (typeof BlockchainNetwork)[keyof typeof BlockchainNetwork];
+export type BlockchainNetwork =
+  (typeof BlockchainNetwork)[keyof typeof BlockchainNetwork];
