+++ lib/common/src/interfaces/index.ts
